+++ magic-unity/Editor/IL2CPPWorkarounds.cs
@@ -8,7 +8,6 @@ using Mono.Cecil.Rocks;
 using UnityEditor;
 using UnityEditor.Build;
 using UnityEngine;
-using System.Reflection;
 
 namespace Magic.Unity
 {
@@ -22,17 +21,15 @@ namespace Magic.Unity
 
         public static void RewriteAssemblies()
         {
-            var cljAssemblies = AppDomain.CurrentDomain
-                            .GetAssemblies()
-                            .Where(a => a.FullName.Contains(".clj"))
-                            .Select(a => a.Location);
-            RewriteAssemblies(cljAssemblies);
+            RewriteAssemblies(PlayerCljAssemblies.Paths());
         }
 
         public static void RewriteAssemblies(IEnumerable<string> files)
         {
+            var fileList = files.ToList();
+            Debug.LogFormat("[Magic.Unity] rewriting {0} clj assemblies", fileList.Count);
             GenerateGenericWorkaroundMethods.Init();
-            foreach (var file in files)
+            foreach (var file in fileList)
             {
                 RewriteAssembly(file);
             }
@@ -40,7 +37,7 @@ namespace Magic.Unity
 
         static void RewriteAssembly(string file)
         {
-            var runtimeLocation = Path.GetDirectoryName(Assembly.Load("Magic.Runtime").Location);
+            var runtimeLocation = PackageExportPath.ExportDirectory;
             Debug.LogFormat($"[Magic.Unity] runtime location {runtimeLocation}");
 
             var resolver = new DefaultAssemblyResolver();